Quadcopter: parameterize yaw as cos,sin projection
This removes the discontinuity at yaw 180deg that was creating
bad behaviors. Given this network is not designed to fly with the
quadcopter inverted, the discontinuity of roll and pitch is not an
issue. We could also work directly with quaternions as the
attitude representation.

This also makes the attitude initialize randomly.